[sdk] update fetchOrders blockchain in SOL
LIQ-912

We couldn't get orders NFT from what's sold in SOL marketplace.

So I temporarily changed the code.
In this case, no matter in devnet or mainnet.
We should use "SOL" when we fetchOrders from backend.(Not use "devnet" or "mainnet-beta")
So I did little mapping change. So that we can get orders.